Add files via upload
Project Background: 
The mass shooting is a bigger concern in the US and current security systems are not equipped to provide the presence of weapons and criminals on the premises. If security securities can identify and notify the presence of suspected criminals in premises, a preventive plan or measures can be taken by authorities to take strategic and operational actions. 
Project Scope: 
Create a computer vision system using AWS services to detect wanted/person of interest and weapon with the help of a Cctv camera.
Services used: S3, Amazon Rekognition, Amazon Sagemaker, IAM